AI_ChatBot/flask_server/requirements.txt
Flask
PyPDF2
langchain
genai
google-generativeai
flask-cors
faiss-gpu
langchain-google-genai
dotenv
csv